Add Validation Module

+++ TriggerSol.JStore/Model/PersistentBase.cs
@@ -32,7 +32,7 @@ using TriggerSol.Dependency;
 
 namespace TriggerSol.JStore
 {
-    public abstract class PersistentBase : NotifyPropertyChangedBase, IPersistentBase, IDependencyObject
+    public abstract class PersistentBase : NotifyPropertyChangedBase, object, IDependencyObject
     {
         protected PersistentBase()
         {
@@ -74,7 +74,7 @@ namespace TriggerSol.JStore
                 DataStore.Save(GetType(), this);
         }
 
-        public IPersistentBase Clone(bool withId = false)
+        public object Clone(bool withId = false)
         {
             var clone = this.CloneObject();
 
@@ -84,12 +84,13 @@ namespace TriggerSol.JStore
             return clone;
         }
 
-        public virtual void Delete()
+        public virtual void Delete(bool allowDeleting = true)
         {
-            DataStore.Delete(GetType(), this);
+            if (allowDeleting)
+                DataStore.Delete(GetType(), this);
         }
 
-        public virtual IPersistentBase Reload()
+        public virtual object Reload()
         {
             return MappingId == null ? null : DataStore.Load(GetType(), MappingId);
         }
@@ -110,19 +111,19 @@ namespace TriggerSol.JStore
             }
         }
 
-        public virtual IList<T> GetAssociatedCollection<T>(string associatedProperty) where T: IPersistentBase
+        public virtual IList<T> GetAssociatedCollection<T>(string associatedProperty) where T: object
         {
             var pInfo = GetProperty(typeof(T).GetTypeInfo(), associatedProperty);
 
             if (pInfo != null)
             {
-                return DataStore.LoadAll<T>(p => pInfo.GetValue(p) != null && ((IPersistentBase)pInfo.GetValue(p)).MappingId != null && ((IPersistentBase)pInfo.GetValue(p)).MappingId.Equals(MappingId)).ToList();
+                return DataStore.LoadAll<T>(p => pInfo.GetValue(p) != null && ((object)pInfo.GetValue(p)).MappingId != null && ((object)pInfo.GetValue(p)).MappingId.Equals(MappingId)).ToList();
             }
 
             return Enumerable.Empty<T>().ToList();
         }
 
-        PropertyInfo GetProperty(TypeInfo typeInfo, string propertyName)
+        protected virtual PropertyInfo GetProperty(TypeInfo typeInfo, string propertyName)
         {
             var pInfo = typeInfo.GetDeclaredProperty(propertyName);
